refactor: relocate animate suggested follows
