+++ src/screens/Profile/Header/SuggestedFollows.tsx
@@ -1,4 +1,11 @@
+import {useEffect, useState} from 'react'
 import {View} from 'react-native'
+import Animated, {
+  Easing,
+  useAnimatedStyle,
+  useSharedValue,
+  withTiming,
+} from 'react-native-reanimated'
 
 import {useSuggestedFollowsByActorQuery} from '#/state/queries/suggested-follows'
 import {atoms as a, useTheme, type ViewStyleProp} from '#/alf'
@@ -67,5 +74,32 @@ export function ProfileHeaderSuggestedFollows({actorDid}: {actorDid: string}) {
       error={error}
       viewContext="profileHeader"
     />
+  )
+}
+
+export const AnimatedSuggestedFollows = ({actorDid}: {actorDid: string}) => {
+  const [contentHeight, setContentHeight] = useState(0)
+  const animatedHeight = useSharedValue(0)
+
+  useEffect(() => {
+    if (contentHeight > 0) {
+      animatedHeight.value = withTiming(contentHeight, {
+        duration: 300,
+        easing: Easing.inOut(Easing.cubic),
+      })
+    }
+  }, [contentHeight, animatedHeight])
+
+  const animatedStyle = useAnimatedStyle(() => ({
+    height: animatedHeight.value,
+    overflow: 'hidden',
+  }))
+
+  return (
+    <Animated.View style={animatedStyle}>
+      <View onLayout={e => setContentHeight(e.nativeEvent.layout.height)}>
+        <ProfileHeaderSuggestedFollows actorDid={actorDid} />
+      </View>
+    </Animated.View>
   )
 }
